Move class RigWellResultBranch into separate file
