+++ ApplicationLibCode/ModelVisualization/RivSimWellPipesPartMgr.cpp
@@ -31,7 +31,6 @@
 #include "RigWellLogExtractor.h"
 #include "RigWellPath.h"
 #include "RigWellResultFrame.h"
-#include "RigWellResultPoint.h"
 
 #include "Rim3dView.h"
 #include "RimCase.h"
