+++ ApplicationLibCode/ReservoirDataModel/RigMswCenterLineCalculator.h
@@ -18,7 +18,7 @@
 
 #pragma once
 
-#include "RigWellResultPoint.h"
+#include "RigWellResultBranch.h"
 
 #include "cvfVector3.h"
