+++ ApplicationLibCode/ReservoirDataModel/RigWellResultPoint.cpp
@@ -301,51 +301,3 @@ cvf::Vec3d RigWellResultPoint::bottomPosition() const
 {
     return m_bottomPosition;
 }
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-int RigWellResultBranch::ertBranchId() const
-{
-    return m_ertBranchId;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RigWellResultBranch::setErtBranchId( int id )
-{
-    m_ertBranchId = id;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-std::vector<RigWellResultPoint> RigWellResultBranch::branchResultPoints() const
-{
-    return m_branchResultPoints;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RigWellResultBranch::clearBranchResultPoints()
-{
-    m_branchResultPoints.clear();
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RigWellResultBranch::addBranchResultPoint( const RigWellResultPoint& point )
-{
-    m_branchResultPoints.push_back( point );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RigWellResultBranch::setBranchResultPoints( const std::vector<RigWellResultPoint>& points )
-{
-    m_branchResultPoints = points;
-}
