+++ ApplicationLibCode/ReservoirDataModel/RigWellResultPoint.h
@@ -90,29 +90,3 @@ private:
     double m_connectionFactor;
     bool   m_isConnectedToValve;
 };
-
-//==================================================================================================
-/// This class contains the connection information from and including a splitpoint to the end of
-/// that particular branch.
-//==================================================================================================
-struct RigWellResultBranch
-{
-    RigWellResultBranch()
-        : m_ertBranchId( -1 )
-    {
-    }
-
-    int  ertBranchId() const;
-    void setErtBranchId( int id );
-
-    std::vector<RigWellResultPoint> branchResultPoints() const;
-    void                            clearBranchResultPoints();
-    void                            addBranchResultPoint( const RigWellResultPoint& point );
-    void                            setBranchResultPoints( const std::vector<RigWellResultPoint>& points );
-
-private:
-    int                             m_ertBranchId;
-    std::vector<RigWellResultPoint> m_branchResultPoints;
-};
-
-using SimulationWellCellBranch = std::pair<std::vector<cvf::Vec3d>, std::vector<RigWellResultPoint>>;
